feat: implement user management features with user creation modal, filters, and user table
