+++ src/app/(admin)/admin/manage-users/utils/constants.js
@@ -1,0 +1,2 @@
+export const USER_TYPES = ["user", "agent", "semi-agent", "sales-inhouse"];
+export const PER_PAGE_OPTIONS = [10, 20, 50, 100];
